[VUMM-896] ModelDB "token has expired" when checking S3
